move TestPipelineUniqueness to PPS tests
