+++ src/server/pachyderm_test.go
@@ -793,550 +793,6 @@ func TestMultipleInputsFromTheSameBranch(t *testing.T) {
 	require.Equal(t, 4, len(commitInfos))
 }
 
-func TestRunPipeline(t *testing.T) {
-	// TODO(2.0 optional): Run pipeline creates a dangling commit, but uses the stats branch for stats commits.
-	// Since there is no relationship between the commits created by run pipeline, there should be no
-	// relationship between the stats commits. So, the stats commits should also be dangling commits.
-	// This might be easier to address when global IDs are implemented.
-	t.Skip("Run pipeline does not work correctly with stats enabled")
-	//if testing.Short() {
-	//	t.Skip("Skipping integration tests in short mode")
-	//}
-
-	//c := tu.GetPachClient(t)
-	//require.NoError(t, c.DeleteAll())
-
-	//// Test on cross pipeline
-	//t.Run("RunPipelineCross", func(t *testing.T) {
-	//	dataRepo := tu.UniqueString("TestRunPipeline_data")
-	//	require.NoError(t, c.CreateProjectRepo(pfs.DefaultProjectName,dataRepo))
-
-	//	branchA := "branchA"
-	//	branchB := "branchB"
-
-	//	pipeline := tu.UniqueString("pipeline")
-	//	require.NoError(t, c.CreateProjectPipeline(pfs.DefaultProjectName,
-	//		pipeline,
-	//		"",
-	//		[]string{"bash"},
-	//		[]string{
-	//			"cat /pfs/branch-a/file >> /pfs/out/file",
-	//			"cat /pfs/branch-b/file >> /pfs/out/file",
-	//			"echo ran-pipeline",
-	//		},
-	//		nil,
-	//		client.NewCrossInput(
-	//			client.NewProjectPFSInputOpts("branch-a",pfs.DefaultProjectName, dataRepo, branchA, "/*", "", "", false, false, nil),
-	//			client.NewProjectPFSInputOpts("branch-b",pfs.DefaultProjectName, dataRepo, branchB, "/*", "", "", false, false, nil),
-	//		),
-	//		"",
-	//		false,
-	//	))
-
-	//	commitA, err := c.StartProjectCommit(pfs.DefaultProjectName,dataRepo, branchA)
-	//	require.NoError(t, err)
-	//	require.NoError(t, c.PutFile(dataRepo, commitA.Branch.Name, commitA.ID, "/file", strings.NewReader("data A\n"), client.WithAppendPutFile()))
-	//	require.NoError(t, c.FinishProjectCommit(pfs.DefaultProjectName,dataRepo, commitA.Branch.Name, commitA.ID))
-
-	//	commitB, err := c.StartProjectCommit(pfs.DefaultProjectName,dataRepo, branchB)
-	//	require.NoError(t, err)
-	//	require.NoError(t, c.PutFile(dataRepo, commitB.Branch.Name, commitB.ID, "/file", strings.NewReader("data B\n"), client.WithAppendPutFile()))
-	//	require.NoError(t, c.FinishProjectCommit(pfs.DefaultProjectName,dataRepo, commitB.Branch.Name, commitB.ID))
-
-	//	iter, err := c.FlushJob([]*pfs.Commit{commitA, commitB}, nil)
-	//	require.NoError(t, err)
-	//	require.Equal(t, 2, len(commits))
-	//	buffer := bytes.Buffer{}
-	//	require.NoError(t, c.GetFile(commits[0].Commit.Repo.Name, commits[0].Commit.Branch.Name, commits[0].Commit.ID, "file", &buffer))
-	//	require.Equal(t, "data A\ndata B\n", buffer.String())
-
-	//	commitM, err := c.StartProjectCommit(pfs.DefaultProjectName,dataRepo, "master")
-	//	require.NoError(t, err)
-	//	require.NoError(t, c.FinishProjectCommit(pfs.DefaultProjectName,dataRepo, commitM.Branch.Name, commitM.ID))
-
-	//	// we should have two jobs
-	//	ji, err := c.ListProjectJob(pfs.DefaultProjectName,pipeline, nil, nil, -1, true)
-	//	require.NoError(t, err)
-	//	require.Equal(t, 2, len(ji))
-	//	// now run the pipeline
-	//	require.NoError(t, c.RunPipeline(pipeline, nil, ""))
-	//	// running the pipeline should create a new job
-	//	require.NoError(t, backoff.Retry(func() error {
-	//		jobInfos, err := c.ListProjectJob(pfs.DefaultProjectName,pipeline, nil, nil, -1, true)
-	//		require.NoError(t, err)
-	//		if len(jobInfos) != 3 {
-	//			return errors.Errorf("expected 3 jobs, got %d", len(jobInfos))
-	//		}
-	//		return nil
-	//	}, backoff.NewTestingBackOff()))
-
-	//	// now run the pipeline with non-empty provenance
-	//	require.NoError(t, backoff.Retry(func() error {
-	//		return c.RunPipeline(pipeline, []*pfs.Commit{
-	//			client.NewProjectCommit(pfs.DefaultProjectName,dataRepo, "branchA", commitA.ID),
-	//		}, "")
-	//	}, backoff.NewTestingBackOff()))
-
-	//	// running the pipeline should create a new job
-	//	require.NoError(t, backoff.Retry(func() error {
-	//		jobInfos, err := c.ListProjectJob(pfs.DefaultProjectName,pipeline, nil, nil, -1, true)
-	//		require.NoError(t, err)
-	//		if len(jobInfos) != 4 {
-	//			return errors.Errorf("expected 4 jobs, got %d", len(jobInfos))
-	//		}
-	//		return nil
-	//	}, backoff.NewTestingBackOff()))
-
-	//	// add some new commits with some new info
-	//	commitA2, err := c.StartProjectCommit(pfs.DefaultProjectName,dataRepo, branchA)
-	//	require.NoError(t, err)
-	//	require.NoError(t, c.PutFile(dataRepo, commitA2.Branch.Name, commitA2.ID, "/file", strings.NewReader("data A2\n"), client.WithAppendPutFile()))
-	//	require.NoError(t, c.FinishProjectCommit(pfs.DefaultProjectName,dataRepo, commitA2.Branch.Name, commitA2.ID))
-
-	//	commitB2, err := c.StartProjectCommit(pfs.DefaultProjectName,dataRepo, branchB)
-	//	require.NoError(t, err)
-	//	require.NoError(t, c.PutFile(dataRepo, commitB2.Branch.Name, commitB2.ID, "/file", strings.NewReader("data B2\n"), client.WithAppendPutFile()))
-	//	require.NoError(t, c.FinishProjectCommit(pfs.DefaultProjectName,dataRepo, commitB2.Branch.Name, commitB2.ID))
-
-	//	// and make sure the output file is updated appropriately
-	//	iter, err = c.FlushJob([]*pfs.Commit{commitA2, commitB2}, nil)
-	//	require.NoError(t, err)
-	//	require.Equal(t, 2, len(commits))
-	//	buffer = bytes.Buffer{}
-	//	require.NoError(t, c.GetFile(commits[0].Commit.Repo.Name, commits[0].Commit.Branch.Name, commits[0].Commit.ID, "file", &buffer))
-	//	require.Equal(t, "data A\ndata A2\ndata B\ndata B2\n", buffer.String())
-
-	//	// now run the pipeline provenant on the old commits
-	//	require.NoError(t, c.RunPipeline(pipeline, []*pfs.Commit{
-	//		client.NewProjectCommit(pfs.DefaultProjectName,dataRepo, "branchA", commitA.ID),
-	//		client.NewProjectCommit(pfs.DefaultProjectName,dataRepo, "branchB", commitB2.ID),
-	//	}, ""))
-
-	//	// and ensure that the file now has the info from the correct versions of the commits
-	//	iter, err = c.FlushJob([]*pfs.Commit{commitA, commitB2}, nil)
-	//	require.NoError(t, err)
-	//	require.Equal(t, 2, len(commits))
-	//	buffer = bytes.Buffer{}
-	//	require.NoError(t, c.GetFile(commits[0].Commit.Repo.Name, commits[0].Commit.Branch.Name, commits[0].Commit.ID, "file", &buffer))
-	//	require.Equal(t, "data A\ndata B\ndata B2\n", buffer.String())
-
-	//	// make sure no commits with this provenance combination exist
-	//	iter, err = c.FlushJob([]*pfs.Commit{commitA2, commitB}, nil)
-	//	require.NoError(t, err)
-	//	require.Equal(t, 0, len(commits))
-	//})
-
-	//// Test on pipeline with no commits
-	//t.Run("RunPipelineEmpty", func(t *testing.T) {
-	//	dataRepo := tu.UniqueString("TestRunPipeline_data")
-	//	require.NoError(t, c.CreateProjectRepo(pfs.DefaultProjectName,dataRepo))
-
-	//	pipeline := tu.UniqueString("empty-pipeline")
-	//	require.NoError(t, c.CreateProjectPipeline(pfs.DefaultProjectName,
-	//		pipeline,
-	//		"",
-	//		nil,
-	//		nil,
-	//		nil,
-	//		nil,
-	//		"",
-	//		false,
-	//	))
-
-	//	// we should have two jobs
-	//	ji, err := c.ListProjectJob(pfs.DefaultProjectName,pipeline, nil, nil, -1, true)
-	//	require.NoError(t, err)
-	//	require.Equal(t, 0, len(ji))
-	//	// now run the pipeline
-	//	require.YesError(t, c.RunPipeline(pipeline, nil, ""))
-	//})
-
-	//// Test on unrelated branch
-	//t.Run("RunPipelineUnrelated", func(t *testing.T) {
-	//	dataRepo := tu.UniqueString("TestRunPipeline_data")
-	//	require.NoError(t, c.CreateProjectRepo(pfs.DefaultProjectName,dataRepo))
-
-	//	branchA := "branchA"
-	//	branchB := "branchB"
-
-	//	pipeline := tu.UniqueString("unrelated-pipeline")
-	//	require.NoError(t, c.CreateProjectPipeline(pfs.DefaultProjectName,
-	//		pipeline,
-	//		"",
-	//		[]string{"bash"},
-	//		[]string{
-	//			"cat /pfs/branch-a/file >> /pfs/out/file",
-	//			"cat /pfs/branch-b/file >> /pfs/out/file",
-	//			"echo ran-pipeline",
-	//		},
-	//		nil,
-	//		client.NewCrossInput(
-	//			client.NewProjectPFSInputOpts("branch-a","", dataRepo, branchA, "/*", "", "", false, false, nil),
-	//			client.NewProjectPFSInputOpts("branch-b","", dataRepo, branchB, "/*", "", "", false, false, nil),
-	//		),
-	//		"",
-	//		false,
-	//	))
-	//	commitA, err := c.StartProjectCommit(pfs.DefaultProjectName,dataRepo, branchA)
-	//	require.NoError(t, err)
-	//	c.PutFile(dataRepo, commitA.Branch.Name, commitA.ID, "/file", strings.NewReader("data A\n", client.WithAppendPutFile()))
-	//	c.FinishProjectCommit(pfs.DefaultProjectName,dataRepo, commitA.Branch.Name, commitA.ID)
-
-	//	commitM, err := c.StartProjectCommit(pfs.DefaultProjectName,dataRepo, "master")
-	//	require.NoError(t, err)
-	//	err = c.FinishProjectCommit(pfs.DefaultProjectName,dataRepo, commitM.Branch.Name, commitM.ID)
-	//	require.NoError(t, err)
-
-	//	require.NoError(t, c.CreateProjectBranch(pfs.DefaultProjectName,dataRepo, "unrelated", "", nil))
-	//	commitU, err := c.StartProjectCommit(pfs.DefaultProjectName,dataRepo, "unrelated")
-	//	require.NoError(t, err)
-	//	err = c.FinishProjectCommit(pfs.DefaultProjectName,dataRepo, commitU.Branch.Name, commitU.ID)
-	//	require.NoError(t, err)
-
-	//	_, err = c.FlushJob([]*pfs.Commit{commitA, commitM, commitU}, nil)
-	//	require.NoError(t, err)
-
-	//	// now run the pipeline with unrelated provenance
-	//	require.YesError(t, c.RunPipeline(pipeline, []*pfs.Commit{
-	//		client.NewProjectCommit(pfs.DefaultProjectName,dataRepo, "unrelated", commitU.ID)}, ""))
-	//})
-
-	//// Test with downstream pipeline
-	//t.Run("RunPipelineDownstream", func(t *testing.T) {
-	//	dataRepo := tu.UniqueString("TestRunPipeline_data")
-	//	require.NoError(t, c.CreateProjectRepo(pfs.DefaultProjectName,dataRepo))
-
-	//	branchA := "branchA"
-	//	branchB := "branchB"
-
-	//	pipeline := tu.UniqueString("original-pipeline")
-	//	require.NoError(t, c.CreateProjectPipeline(pfs.DefaultProjectName,
-	//		pipeline,
-	//		"",
-	//		[]string{"bash"},
-	//		[]string{
-	//			"cat /pfs/branch-a/file >> /pfs/out/file",
-	//			"cat /pfs/branch-b/file >> /pfs/out/file",
-	//			"echo ran-pipeline",
-	//		},
-	//		nil,
-	//		client.NewCrossInput(
-	//			client.NewProjectPFSInputOpts("branch-a","", dataRepo, branchA, "/*", "", "", false, false, nil),
-	//			client.NewProjectPFSInputOpts("branch-b","", dataRepo, branchB, "/*", "", "", false, false, nil),
-	//		),
-	//		"",
-	//		false,
-	//	))
-
-	//	commitA, err := c.StartProjectCommit(pfs.DefaultProjectName,dataRepo, branchA)
-	//	require.NoError(t, err)
-	//	c.PutFile(dataRepo, commitA.Branch.Name, commitA.ID, "/file", strings.NewReader("data A\n", client.WithAppendPutFile()))
-	//	c.FinishProjectCommit(pfs.DefaultProjectName,dataRepo, commitA.Branch.Name, commitA.ID)
-
-	//	commitB, err := c.StartProjectCommit(pfs.DefaultProjectName,dataRepo, branchB)
-	//	require.NoError(t, err)
-	//	c.PutFile(dataRepo, commitB.Branch.Name, commitB.ID, "/file", strings.NewReader("data B\n", client.WithAppendPutFile()))
-	//	c.FinishProjectCommit(pfs.DefaultProjectName,dataRepo, commitB.Branch.Name, commitB.ID)
-
-	//	iter, err := c.FlushJob([]*pfs.Commit{commitA, commitB}, nil)
-	//	require.NoError(t, err)
-	//	require.Equal(t, 2, len(commits))
-	//	buffer := bytes.Buffer{}
-	//	require.NoError(t, c.GetFile(commits[0].Commit.Repo.Name, commits[0].Commit.Branch.Name, commits[0].Commit.ID, "file", &buffer))
-	//	require.Equal(t, "data A\ndata B\n", buffer.String())
-
-	//	// and make sure we can attatch a downstream pipeline
-	//	downstreamPipeline := tu.UniqueString("downstream-pipeline")
-	//	require.NoError(t, c.CreateProjectPipeline(pfs.DefaultProjectName,
-	//		downstreamPipeline,
-	//		"",
-	//		[]string{"/bin/bash"},
-	//		[]string{"cp " + fmt.Sprintf("/pfs/%s/*", pipeline) + " /pfs/out/"},
-	//		nil,
-	//		client.NewProjectPFSInput(pfs.DefaultProjectName,pipeline, "/*"),
-	//		"",
-	//		false,
-	//	))
-
-	//	commitA2, err := c.StartProjectCommit(pfs.DefaultProjectName,dataRepo, branchA)
-	//	require.NoError(t, err)
-	//	err = c.FinishProjectCommit(pfs.DefaultProjectName,dataRepo, commitA2.Branch.Name, commitA2.ID)
-	//	require.NoError(t, err)
-
-	//	// there should be one job on the old commit for downstreamPipeline
-	//	jobInfos, err := c.FlushJobAll([]*pfs.Commit{commitA}, []string{downstreamPipeline})
-	//	require.NoError(t, err)
-	//	require.Equal(t, 1, len(jobInfos))
-
-	//	// now run the pipeline
-	//	require.NoError(t, backoff.Retry(func() error {
-	//		return c.RunPipeline(pipeline, []*pfs.Commit{
-	//			client.NewProjectCommit(pfs.DefaultProjectName,dataRepo, branchA, commitA.ID),
-	//		}, "")
-	//	}, backoff.NewTestingBackOff()))
-
-	//	// the downstream pipeline shouldn't have any new jobs, since runpipeline jobs don't propagate
-	//	jobInfos, err = c.FlushJobAll([]*pfs.Commit{commitA}, []string{downstreamPipeline})
-	//	require.NoError(t, err)
-	//	require.Equal(t, 1, len(jobInfos))
-
-	//	// now rerun the one job that we saw
-	//	require.NoError(t, backoff.Retry(func() error {
-	//		return c.RunPipeline(downstreamPipeline, nil, jobInfos[0].Job.ID)
-	//	}, backoff.NewTestingBackOff()))
-
-	//	// we should now have two jobs
-	//	jobInfos, err = c.FlushJobAll([]*pfs.Commit{commitA}, []string{downstreamPipeline})
-	//	require.NoError(t, err)
-	//	require.Equal(t, 2, len(jobInfos))
-	//})
-
-	//// Test with a downstream pipeline who's upstream has no datum, but where the downstream still needs to succeed
-	//t.Run("RunPipelineEmptyUpstream", func(t *testing.T) {
-	//	dataRepo := tu.UniqueString("TestRunPipeline_data")
-	//	require.NoError(t, c.CreateProjectRepo(pfs.DefaultProjectName,dataRepo))
-
-	//	branchA := "branchA"
-	//	branchB := "branchB"
-
-	//	pipeline := tu.UniqueString("pipeline-downstream")
-	//	require.NoError(t, c.CreateProjectPipeline(pfs.DefaultProjectName,
-	//		pipeline,
-	//		"",
-	//		[]string{"bash"},
-	//		[]string{
-	//			"cat /pfs/branch-a/file >> /pfs/out/file",
-	//			"cat /pfs/branch-b/file >> /pfs/out/file",
-	//			"echo ran-pipeline",
-	//		},
-	//		nil,
-	//		client.NewCrossInput(
-	//			client.NewProjectPFSInputOpts("branch-a","", dataRepo, branchA, "/*", "", "", false, false, nil),
-	//			client.NewProjectPFSInputOpts("branch-b","", dataRepo, branchB, "/*", "", "", false, false, nil),
-	//		),
-	//		"",
-	//		false,
-	//	))
-
-	//	commitA, err := c.StartProjectCommit(pfs.DefaultProjectName,dataRepo, branchA)
-	//	require.NoError(t, err)
-	//	c.PutFile(dataRepo, commitA.Branch.Name, commitA.ID, "/file", strings.NewReader("data A\n", client.WithAppendPutFile()))
-	//	c.FinishProjectCommit(pfs.DefaultProjectName,dataRepo, commitA.Branch.Name, commitA.ID)
-
-	//	iter, err := c.FlushJob([]*pfs.Commit{commitA}, nil)
-	//	require.NoError(t, err)
-	//	require.Equal(t, 2, len(commits))
-
-	//	// no commit to branch-b so "file" should not exist
-	//	buffer := bytes.Buffer{}
-	//	require.YesError(t, c.GetFile(commits[0].Commit.Repo.Name, commits[0].Commit.Branch.Name, commits[0].Commit.ID, "file", &buffer))
-
-	//	// and make sure we can attatch a downstream pipeline
-	//	downstreamPipeline := tu.UniqueString("pipelinedownstream")
-	//	require.NoError(t, c.CreateProjectPipeline(pfs.DefaultProjectName,
-	//		downstreamPipeline,
-	//		"",
-	//		[]string{"/bin/bash"},
-	//		[]string{
-	//			"cat /pfs/branch-a/file >> /pfs/out/file",
-	//			fmt.Sprintf("cat /pfs/%s/file >> /pfs/out/file", pipeline),
-	//			"echo ran-pipeline",
-	//		},
-	//		nil,
-	//		client.NewUnionInput(
-	//			client.NewProjectPFSInputOpts("branch-a","", dataRepo, branchA, "/*", "", "", false, false, nil),
-	//			client.NewProjectPFSInput(pfs.DefaultProjectName,pipeline, "/*"),
-	//		),
-	//		"",
-	//		false,
-	//	))
-
-	//	commitA2, err := c.StartProjectCommit(pfs.DefaultProjectName,dataRepo, branchA)
-	//	require.NoError(t, err)
-	//	err = c.FinishProjectCommit(pfs.DefaultProjectName,dataRepo, commitA2.Branch.Name, commitA2.ID)
-	//	require.NoError(t, err)
-
-	//	// there should be one job on the old commit for downstreamPipeline
-	//	jobInfos, err := c.FlushJobAll([]*pfs.Commit{commitA}, []string{downstreamPipeline})
-	//	require.NoError(t, err)
-	//	require.Equal(t, 1, len(jobInfos))
-
-	//	// now run the pipeline
-	//	require.NoError(t, backoff.Retry(func() error {
-	//		return c.RunPipeline(pipeline, []*pfs.Commit{
-	//			client.NewProjectCommit(pfs.DefaultProjectName,dataRepo, branchA, commitA.ID),
-	//		}, "")
-	//	}, backoff.NewTestingBackOff()))
-
-	//	buffer2 := bytes.Buffer{}
-	//	require.NoError(t, c.GetFile(jobInfos[0].OutputCommit.Repo.Name, jobInfos[0].OutputCommit.Branch.Name, jobInfos[0].OutputCommit.ID, "file", &buffer2))
-	//	// the union of an empty output and datA should only return a file with "data A" in it.
-	//	require.Equal(t, "data A\n", buffer2.String())
-
-	//	// add another commit to see that we can successfully do the cross and union together
-	//	commitB, err := c.StartProjectCommit(pfs.DefaultProjectName,dataRepo, branchB)
-	//	require.NoError(t, err)
-	//	c.PutFile(dataRepo, commitB.Branch.Name, commitB.ID, "/file", strings.NewReader("data B\n", client.WithAppendPutFile()))
-	//	c.FinishProjectCommit(pfs.DefaultProjectName,dataRepo, commitB.Branch.Name, commitB.ID)
-
-	//	_, err = c.FlushJob([]*pfs.Commit{commitA, commitB}, nil)
-	//	require.NoError(t, err)
-
-	//	jobInfos, err = c.FlushJobAll([]*pfs.Commit{commitB}, []string{downstreamPipeline})
-	//	require.NoError(t, err)
-	//	require.Equal(t, 1, len(jobInfos))
-
-	//	buffer3 := bytes.Buffer{}
-	//	require.NoError(t, c.GetFile(jobInfos[0].OutputCommit.Repo.Name, jobInfos[0].OutputCommit.Branch.Name, jobInfos[0].OutputCommit.ID, "file", &buffer3))
-	//	// now that we've added data to the other branch of the cross, we should see the union of data A along with the the crossed data.
-	//	require.Equal(t, "data A\ndata A\ndata B\n", buffer3.String())
-	//})
-
-	//// Test on commits from the same branch
-	//t.Run("RunPipelineSameBranch", func(t *testing.T) {
-	//	dataRepo := tu.UniqueString("TestRunPipeline_data")
-	//	require.NoError(t, c.CreateProjectRepo(pfs.DefaultProjectName,dataRepo))
-
-	//	branchA := "branchA"
-	//	branchB := "branchB"
-
-	//	pipeline := tu.UniqueString("sameBranch-pipeline")
-	//	require.NoError(t, c.CreateProjectPipeline(pfs.DefaultProjectName,
-	//		pipeline,
-	//		"",
-	//		[]string{"bash"},
-	//		[]string{
-	//			"cat /pfs/branch-a/file >> /pfs/out/file",
-	//			"cat /pfs/branch-b/file >> /pfs/out/file",
-	//			"echo ran-pipeline",
-	//		},
-	//		nil,
-	//		client.NewCrossInput(
-	//			client.NewProjectPFSInputOpts("branch-a","", dataRepo, branchA, "/*", "", "", false, false, nil),
-	//			client.NewProjectPFSInputOpts("branch-b","", dataRepo, branchB, "/*", "", "", false, false, nil),
-	//		),
-	//		"",
-	//		false,
-	//	))
-	//	commitA1, err := c.StartProjectCommit(pfs.DefaultProjectName,dataRepo, branchA)
-	//	require.NoError(t, err)
-	//	c.PutFile(dataRepo, commitA1.Branch.Name, commitA1.ID, "/file", strings.NewReader("data A1\n", client.WithAppendPutFile()))
-	//	c.FinishProjectCommit(pfs.DefaultProjectName,dataRepo, commitA1.Branch.Name, commitA1.ID)
-
-	//	commitA2, err := c.StartProjectCommit(pfs.DefaultProjectName,dataRepo, branchA)
-	//	require.NoError(t, err)
-	//	c.PutFile(dataRepo, commitA2.Branch.Name, commitA2.ID, "/file", strings.NewReader("data A2\n", client.WithAppendPutFile()))
-	//	c.FinishProjectCommit(pfs.DefaultProjectName,dataRepo, commitA2.Branch.Name, commitA2.ID)
-
-	//	_, err = c.FlushJob([]*pfs.Commit{commitA1, commitA2}, nil)
-	//	require.NoError(t, err)
-
-	//	// now run the pipeline with provenance from the same branch
-	//	require.YesError(t, c.RunPipeline(pipeline, []*pfs.Commit{
-	//		client.NewProjectCommit(pfs.DefaultProjectName,dataRepo, branchA, commitA1.ID),
-	//		client.NewProjectCommit(pfs.DefaultProjectName,dataRepo, branchA, commitA2.ID),
-	//	}, ""))
-	//})
-	//// Test on pipeline that should always fail
-	//t.Run("RerunPipeline", func(t *testing.T) {
-	//	dataRepo := tu.UniqueString("TestRerunPipeline_data")
-	//	require.NoError(t, c.CreateProjectRepo(pfs.DefaultProjectName,dataRepo))
-
-	//	// jobs on this pipeline should always fail
-	//	pipeline := tu.UniqueString("rerun-pipeline")
-	//	require.NoError(t, c.CreateProjectPipeline(pfs.DefaultProjectName,
-	//		pipeline,
-	//		"",
-	//		[]string{"bash"},
-	//		[]string{"false"},
-	//		nil,
-	//		client.NewProjectPFSInputOpts("branch-a","", dataRepo, "branchA", "/*", "", "", false, false, nil),
-	//		"",
-	//		false,
-	//	))
-
-	//	commitA1, err := c.StartProjectCommit(pfs.DefaultProjectName,dataRepo, "branchA")
-	//	require.NoError(t, err)
-	//	require.NoError(t, c.PutFile(dataRepo, commitA1.Branch.Name, commitA1.ID, "/file", strings.NewReader("data A1\n"), client.WithAppendPutFile()))
-	//	require.NoError(t, c.FinishProjectCommit(pfs.DefaultProjectName,dataRepo, commitA1.Branch.Name, commitA1.ID))
-
-	//	iter, err := c.FlushJob([]*pfs.Commit{commitA1}, nil)
-	//	require.NoError(t, err)
-	//	require.Equal(t, 2, len(commits))
-	//	// now run the pipeline
-	//	require.NoError(t, c.RunPipeline(pipeline, nil, ""))
-
-	//	// running the pipeline should create a new job
-	//	require.NoError(t, backoff.Retry(func() error {
-	//		jobInfos, err := c.ListProjectJob(pfs.DefaultProjectName,pipeline, nil, nil, -1, true)
-	//		require.NoError(t, err)
-	//		if len(jobInfos) != 2 {
-	//			return errors.Errorf("expected 2 jobs, got %d", len(jobInfos))
-	//		}
-
-	//		// but both of these jobs should fail
-	//		for i, job := range jobInfos {
-	//			if job.State.String() != "JOB_FAILURE" {
-	//				return errors.Errorf("expected job %v to fail, but got %v", i, job.State.String())
-	//			}
-	//		}
-	//		return nil
-	//	}, backoff.NewTestingBackOff()))
-
-	//	// Shouldn't error if you try to delete an already deleted pipeline
-	//	require.NoError(t, c.DeleteProjectPipeline(pfs.DefaultProjectName,pipeline, false))
-	//	require.NoError(t, c.DeleteProjectPipeline(pfs.DefaultProjectName,pipeline, false))
-	//})
-	//t.Run("RunPipelineStats", func(t *testing.T) {
-	//	dataRepo := tu.UniqueString("TestRunPipeline_data")
-	//	require.NoError(t, c.CreateProjectRepo(pfs.DefaultProjectName,dataRepo))
-
-	//	branchA := "branchA"
-
-	//	pipeline := tu.UniqueString("stats-pipeline")
-	//	_, err := c.PpsAPIClient.CreatePipeline(
-	//		context.Background(),
-	//		&pps.CreatePipelineRequest{
-	//			Pipeline: client.NewProjectPipeline(pfs.DefaultProjectName,pipeline),
-	//			Transform: &pps.Transform{
-	//				Cmd: []string{"bash"},
-	//				Stdin: []string{
-	//					"cat /pfs/branch-a/file >> /pfs/out/file",
-
-	//					"echo ran-pipeline",
-	//				},
-	//			},
-	//			Input:       client.NewProjectPFSInputOpts("branch-a","", dataRepo, branchA, "/*", "", "", false, false, nil),
-	//		})
-	//	require.NoError(t, err)
-
-	//	commitA, err := c.StartProjectCommit(pfs.DefaultProjectName,dataRepo, branchA)
-	//	require.NoError(t, err)
-	//	c.PutFile(dataRepo, commitA.Branch.Name, commitA.ID, "/file", strings.NewReader("data A\n", client.WithAppendPutFile()))
-	//	c.FinishProjectCommit(pfs.DefaultProjectName,dataRepo, commitA.Branch.Name, commitA.ID)
-
-	//	// wait for the commit to finish before calling RunPipeline
-	//	_, err = c.WaitCommitSetAll([]*pfs.Commit{client.NewProjectCommit(pfs.DefaultProjectName,dataRepo, commitA.ID)}, nil)
-	//	require.NoError(t, err)
-
-	//	// now run the pipeline
-	//	require.NoError(t, backoff.Retry(func() error {
-	//		return c.RunPipeline(pipeline, []*pfs.Commit{
-	//			client.NewProjectCommit(pfs.DefaultProjectName,dataRepo, branchA, commitA.ID),
-	//		}, "")
-	//	}, backoff.NewTestingBackOff()))
-
-	//	// make sure the pipeline didn't crash
-	//	commitInfos, err := c.WaitCommitSetAll([]*pfs.Commit{client.NewProjectCommit(pfs.DefaultProjectName,dataRepo, commitA.ID)}, nil)
-	//	require.NoError(t, err)
-
-	//	// we'll know it crashed if this causes it to hang
-	//	require.NoErrorWithinTRetry(t, 80*time.Second, func() error {
-	//		return nil
-	//	})
-	//})
-}
-
 func TestPipelineJobHasAuthToken(t *testing.T) {
 	if testing.Short() {
 		t.Skip("Skipping integration tests in short mode")
@@ -2870,45 +2326,6 @@ func TestRecursiveCp(t *testing.T) {
 	require.NoError(t, c.FinishCommit(pfs.DefaultProjectName, dataRepo, "", commit.Id))
 	_, err = c.WaitCommitSetAll(commit.Id)
 	require.NoError(t, err)
-}
-
-func TestPipelineUniqueness(t *testing.T) {
-	if testing.Short() {
-		t.Skip("Skipping integration tests in short mode")
-	}
-
-	t.Parallel()
-	c, _ := minikubetestenv.AcquireCluster(t)
-
-	repo := tu.UniqueString("data")
-	require.NoError(t, c.CreateRepo(pfs.DefaultProjectName, repo))
-	pipelineName := tu.UniqueString("pipeline")
-	require.NoError(t, c.CreatePipeline(pfs.DefaultProjectName,
-		pipelineName,
-		"",
-		[]string{"bash"},
-		[]string{""},
-		&pps.ParallelismSpec{
-			Constant: 1,
-		},
-		client.NewPFSInput(pfs.DefaultProjectName, repo, "/"),
-		"",
-		false,
-	))
-	err := c.CreatePipeline(pfs.DefaultProjectName,
-		pipelineName,
-		"",
-		[]string{"bash"},
-		[]string{""},
-		&pps.ParallelismSpec{
-			Constant: 1,
-		},
-		client.NewPFSInput(pfs.DefaultProjectName, repo, "/"),
-		"",
-		false,
-	)
-	require.YesError(t, err)
-	require.Matches(t, "pipeline .*? already exists", err.Error())
 }
 
 func TestUpdatePipeline(t *testing.T) {
